requirements.txt
absl-py==0.12.0
actionlib==1.13.2
altair==4.1.0
angles==1.9.13
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.2.0
base58==2.1.0
bcrypt==3.1.7
bleach==3.3.0
blinker==1.4
bondpy==1.8.6
Brlapi==0.7.0
CacheControl==0.12.6
cachetools==4.2.1
cachy==0.3.0
camera-calibration==1.15.3
camera-calibration-parsers==1.12.0
catkin==0.8.9
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
cbor==1.0.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
cleo==0.8.1
click==7.1.2
clikit==0.6.2
cloudpickle==1.6.0
cmake==3.18.2.post1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.4
controller-manager-msgs==0.19.4
crashtest==0.3.1
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.15.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diagnostic-analysis==1.10.3
diagnostic-common-diagnostics==1.10.3
diagnostic-updater==1.10.3
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dlib==19.18.0
docutils==0.16
duplicity==0.8.12.0
dynamic-reconfigure==1.7.1
edgetpu==2.15.0
empy==3.3.2
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gazebo-plugins==2.9.1
gazebo-ros==2.9.1
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.14
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.10
image-geometry==1.15.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
imutils==0.5.4
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kazam==1.4.5
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.8.0
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.5.3
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.1.2
message-filters==1.15.9
mistune==0.8.4
monotonic==1.5
moviepy==1.0.3
mpi4py==3.0.3
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
nose==1.3.7
notebook==6.3.0
numpy==1.18.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkginfo==1.7.0
poetry==1.1.5
poetry-core==1.0.2
proglog==0.1.9
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.5.1
ptyprocess==0.7.0
py-ubjson==0.14.0
pyarrow==3.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydeck==0.6.2
pydot==1.4.1
pyglet==1.5.0
Pygments==2.3.1
PyGObject==3.36.0
pygraphviz==1.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pylev==1.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
PyQt5-sip==12.8.1
pyqtgraph==0.12.0
pyRFC3339==1.1
pyrsistent==0.17.3
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
python-qt-binding==0.4.3
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.15.9
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosdep-modules==0.20.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.9
roslaunch==1.15.9
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.9
rosmake==1.15.7
rosmaster==1.15.9
rosmsg==1.15.9
rosnode==1.15.9
rosparam==1.15.9
rospkg==1.2.10
rospkg-modules==1.2.10
rospy==1.15.9
rosservice==1.15.9
rostest==1.15.9
rostopic==1.15.9
rosunit==1.15.7
roswtf==1.15.9
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.10
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.9
rqt-msg==0.4.9
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.10
rqt-publisher==0.4.9
rqt-py-common==0.5.2
rqt-py-console==0.4.9
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.8
rqt-rviz==0.6.1
rqt-service-caller==0.4.9
rqt-shell==0.4.10
rqt-srv==0.4.8
rqt-tf-tree==0.6.2
rqt-top==0.4.9
rqt-topic==0.4.12
rqt-web==0.4.9
rsa==4.7.2
rviz==1.14.5
scipy==1.6.1
screen-resolution-extra==0.0.0
screeninfo==0.6.7
SecretStorage==3.3.1
Send2Trash==1.5.0
sensor-msgs==1.13.1
service-identity==18.1.0
shellingham==1.4.0
simplejson==3.16.0
sip==4.19.21
six==1.15.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
smmap==4.0.0
ssh-import-id==5.10
streamlit==0.80.0
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
toml==0.10.2
tomlkit==0.7.0
toolz==0.11.1
topic-tools==1.15.9
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
Twisted==18.9.0
txaio==2.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
u-msgpack-python==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
usb-creator==0.3.7
validators==0.18.2
virtualenv==20.4.3
wadllib==1.3.3
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
wsaccel==0.6.2
xacro==1.14.6
xkit==0.0.0
zope.interface==4.7.1
